requirements.txt
numpy
pyyaml
bs4
tqdm
kanjize
wordcloud
requests
pandas
ttkthemes
pyinstaller

# 形態素解析ライブラリ
# sudachipy
# sudachidict_small/sdachidict_core/sudachidict_full
# # アプリケーション化としては未対応
# janome

# doc2vec用
# gensim==3.8.3

# jupyter用
# ipython